Refactor BooleanInference module to improve structure and clarity.
- Replace `region_cache.jl` with direct region management in `problems.jl`.
- Update variable and tensor ID handling to use `Int` instead of custom structs.
- Enhance debug logging in various functions for better traceability.
- Refactor function signatures and internal logic for consistency and clarity.

+++ src/refactor/utils.jl
@@ -1,16 +1,16 @@
-@inline function get_unfixed_vars(doms::Vector{DomainMask})::Vector{Int32}
-    unfixed_vars = Int32[]
+@inline function get_unfixed_vars(doms::Vector{DomainMask})::Vector{Int}
+    unfixed_vars = Int[]
     @inbounds for (i, dm) in enumerate(doms)
         if !is_fixed(dm)
-            push!(unfixed_vars, Int32(i))
+            push!(unfixed_vars, i)
         end
     end
     return unfixed_vars
 end
 
 # Convenience: compute number of unfixed variables quickly.
-@inline function count_unfixed(doms::Vector{DomainMask})::Int32
-    c::Int32 = 0
+@inline function count_unfixed(doms::Vector{DomainMask})::Int
+    c::Int = 0
     @inbounds for dm in doms
         if !is_fixed(dm)
             c += 1
